Moving and renaming of files into folders. Creation of landing page.
Just renaming a few files and putting them into specific folders. Made a landing page to load the basic website page I had before.
