+++ src/App.js
@@ -1,14 +1,14 @@
-import './App.css';
+import React from 'react';
+import {Route, Switch} from 'react-router-dom';
+import Landing from './pages/Landing';
 
-function App() {
+const App = () => {
   return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          Here is the start of the project!
-        </p>
-      </header>
-    </div>
+     <Switch>
+       <Route exact path = "/">
+         <Landing />
+        </Route>
+      </Switch>
   );
 }
