+++ src/css/Landing.css
@@ -1,8 +1,8 @@
-.App {
+.Landing {
   text-align: center;
 }
 
-.App-header {
+.Landing-header {
   background-color: #282c34;
   min-height: 100vh;
   display: flex;
