+++ src/pages/Landing.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../css/Landing.css';
+
+const Landing = () => {
+    return (
+        <div className="Landing">
+          <header className="Landing-header">
+            <p>
+              Here is the start of the project!
+            </p>
+          </header>
+        </div>
+      );
+}
+export default Landing;
